src/requirements.txt
django == 2.1.5
djangorestframework == 3.9.0
psycopg2-binary
flower == 0.9.2
Celery == 4.2.1
redis == 2.10.6
django_celery_results == 1.0.1
django-celery-beat == 1.4.0
Jinja2 == 2.10
tornado == 5.1.1
requests == 2.20.0
django-filter == 2.1.0